+++ CesiumUtility/src/Color.cpp
@@ -2,6 +2,7 @@
 
 namespace CesiumUtility {
 uint32_t Color::toRgba32() const {
-  return this->a << 24 | this->r << 16 | this->g << 8 | this->b;
+  return (uint32_t)this->a << 24 | (uint32_t)this->r << 16 |
+         (uint32_t)this->g << 8 | this->b;
 }
 } // namespace CesiumUtility
